+++ example/samples/publisher.py
@@ -1,6 +1,9 @@
 #! /usr/bin/env python
 import datetime
-from couchbasekit import Document
+from couchbasekit import Document, Connection
+from example.samples.book import Book
+
+Connection.auth('couchbasekit_samples', 'couchbasekit')
 
 class Publisher(Document):
     __bucket_name__ = 'couchbasekit_samples'
@@ -13,6 +16,9 @@ class Publisher(Document):
         'address': unicode,
         'established_year': int,
         'created_at': datetime.datetime,
+        'publishes': {
+            datetime.date: [Book],
+        }
     }
     # optional
     default_values = {