+++ example/samples/user.py
@@ -1,7 +1,9 @@
 #! /usr/bin/env python
 import datetime
-from couchbasekit import Document
+from couchbasekit import Document, Connection
 from couchbasekit.fields import EmailField, PasswordField
+
+Connection.auth('couchbasekit_samples', 'couchbasekit')
 
 class User(Document):
     __bucket_name__ = 'couchbasekit_samples'