+++ CHANGELOG.rst
@@ -1,12 +1,12 @@
 Changelog
 =========
 
-Release v0.1.1-dev
+Release v0.1.1
 --------------
 
 * Fix timezone aware datetime/time bug for free list fields
-* Fix model document validation
-* Removed __key_field__ dependency on pragmatic document relations
+* Fix model document validation bugs
+* Removed pragmatic document relations due to complexity
 * Exception single/double quote swap
 * ChoiceField.iteritems() now returns ChoiceField.CHOICES.iteritems()
 * Respect lazy translations in ChoiceField